chore(deps): 更新项目依赖并生成新的Cargo.lock文件
- 添加了新的依赖包及其版本信息到锁文件
- 包含各种库的更新，如aead、aes、base64等
- 记录了所有依赖包的校验和以确保构建一致性
- 更新了依赖关系树和版本兼容性信息

+++ pier-core/src/logging.rs
@@ -215,7 +215,7 @@ fn civil_from_unix(secs: i64) -> (i32, u32, u32, u32, u32, u32) {
     };
     let doe = (z - era * 146_097) as u32; // [0, 146096]
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365; // [0, 399]
-    let y = (yoe as i64 + era * 400) as i64;
+    let y = yoe as i64 + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
     let mp = (5 * doy + 2) / 153; // [0, 11]
     let d = doy - (153 * mp + 2) / 5 + 1; // [1, 31]
